requirements.txt
# requirements.txt
Flask==2.3.3
mysql-connector-python
reportlab==4.0.4
python-dotenv==1.0.0
python-dateutil==2.8.2
Pillow==10.0.0
openai==0.28.1
twilio==8.10.0
gunicorn==21.2.0
cryptography==41.0.4